azfunction/requirements.txt
azure-functions
scikit-learn